+++ src/hooks/ready/registerGameSettings.js
@@ -407,6 +407,14 @@ export default function () {
     default: true,
   });
 
+  game.settings.register("ddb-importer", "munching-policy-hide-description", {
+    name: "ddb-importer.munching-policy-hide-description.name",
+    hint: "ddb-importer.munching-policy-hide-description.hint",
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: false,
+  });
 
   // sets the current tutorial step the user has completed
   // game.settings.register("ddb-importer", "tutorial-step", {